Please create a new controller action, which will open a form The form should contain the following fields: - Referrer email address - Referrer name - Recipient email address - Recipient name - Message to send When a customer submits this form, it should be stored in the database.
Please create a grid in the admin interface, which will show these values and an administrator can edit/delete them.

When you create the module, please place the database creation script in an install script in the module's sql directory.

When you create the module, please place a new entry in an install script in the module's data directory.